rename and first build
